+++ src/blmm_batch.py
